% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudhsm_operations.R
\name{get_config}
\alias{get_config}
\title{This is documentation for AWS CLOUDHSM CLASSIC}
\usage{
get_config(ClientArn, ClientVersion, HapgList)
}
\arguments{
\item{ClientArn}{[required] The ARN of the client.}

\item{ClientVersion}{[required] The client version.}

\item{HapgList}{[required] A list of ARNs that identify the high-availability partition groups that are associated with the client.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more information, see \href{http://aws.amazon.com/cloudhsm/faqs-classic/}{AWS CloudHSM Classic FAQs}, the \href{http://docs.aws.amazon.com/cloudhsm/classic/userguide/}{AWS CloudHSM Classic User Guide}, and the \href{http://docs.aws.amazon.com/cloudhsm/classic/APIReference/}{AWS CloudHSM Classic API Reference}.
}
\details{
\strong{For information about the current version of AWS CloudHSM}, see \href{http://aws.amazon.com/cloudhsm/}{AWS CloudHSM}, the \href{http://docs.aws.amazon.com/cloudhsm/latest/userguide/}{AWS CloudHSM User Guide}, and the \href{http://docs.aws.amazon.com/cloudhsm/latest/APIReference/}{AWS CloudHSM API Reference}.

Gets the configuration files necessary to connect to all high availability partition groups the client is associated with.
}
\section{Accepted Parameters}{
\preformatted{get_config(
  ClientArn = "string",
  ClientVersion = "5.1"|"5.3",
  HapgList = list(
    "string"
  )
)
}
}

