% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudhsm_operations.R
\name{list_hsms}
\alias{list_hsms}
\title{This is documentation for AWS CLOUDHSM CLASSIC}
\usage{
list_hsms(NextToken = NULL)
}
\arguments{
\item{NextToken}{The \code{NextToken} value from a previous call to \code{ListHsms}. Pass null if this is the first call.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more information, see \href{http://aws.amazon.com/cloudhsm/faqs-classic/}{AWS CloudHSM Classic FAQs}, the \href{http://docs.aws.amazon.com/cloudhsm/classic/userguide/}{AWS CloudHSM Classic User Guide}, and the \href{http://docs.aws.amazon.com/cloudhsm/classic/APIReference/}{AWS CloudHSM Classic API Reference}.
}
\details{
\strong{For information about the current version of AWS CloudHSM}, see \href{http://aws.amazon.com/cloudhsm/}{AWS CloudHSM}, the \href{http://docs.aws.amazon.com/cloudhsm/latest/userguide/}{AWS CloudHSM User Guide}, and the \href{http://docs.aws.amazon.com/cloudhsm/latest/APIReference/}{AWS CloudHSM API Reference}.

Retrieves the identifiers of all of the HSMs provisioned for the current customer.

This operation supports pagination with the use of the \code{NextToken} member. If more results are available, the \code{NextToken} member of the response contains a token that you pass in the next call to \code{ListHsms} to retrieve the next set of items.
}
\section{Accepted Parameters}{
\preformatted{list_hsms(
  NextToken = "string"
)
}
}

