% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudhsm_operations.R
\name{modify_hapg}
\alias{modify_hapg}
\title{This is documentation for \strong{AWS CloudHSM Classic}}
\usage{
modify_hapg(HapgArn, Label = NULL, PartitionSerialList = NULL)
}
\arguments{
\item{HapgArn}{The ARN of the high-availability partition group to modify.}

\item{Label}{The new label for the high-availability partition group.}

\item{PartitionSerialList}{The list of partition serial numbers to make members of the high-availability partition group.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more information, see \href{http://aws.amazon.com/cloudhsm/faqs-classic/}{AWS CloudHSM Classic FAQs}, the \href{http://docs.aws.amazon.com/cloudhsm/classic/userguide/}{AWS CloudHSM Classic User Guide}, and the \href{http://docs.aws.amazon.com/cloudhsm/classic/APIReference/}{AWS CloudHSM Classic API Reference}.
}
\details{
\strong{For information about the current version of AWS CloudHSM}, see \href{http://aws.amazon.com/cloudhsm/}{AWS CloudHSM}, the \href{http://docs.aws.amazon.com/cloudhsm/latest/userguide/}{AWS CloudHSM User Guide}, and the \href{http://docs.aws.amazon.com/cloudhsm/latest/APIReference/}{AWS CloudHSM API Reference}.

Modifies an existing high-availability partition group.
}
\examples{

}
