% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloudhsm_package.R
\docType{package}
\name{paws.cloudhsm-package}
\alias{paws.cloudhsm}
\alias{paws.cloudhsm-package}
\title{paws.cloudhsm: Amazon CloudHSM}
\description{
This is documentation for **AWS CloudHSM Classic**.
For more information, see [AWS CloudHSM Classic
FAQs](http://aws.amazon.com/cloudhsm/faqs-classic/), the [AWS CloudHSM
Classic User
Guide](http://docs.aws.amazon.com/cloudhsm/classic/userguide/), and
the [AWS CloudHSM Classic API
Reference](http://docs.aws.amazon.com/cloudhsm/classic/APIReference/).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
